feat: remove authorizations when device is added
When a device is (re-)added we have to ensure the authorized service
providers are removed. When there is a new owner of a device the old
authorizations need to be cleaned.

Note that in this case the secret is still the same so in theory the old
owner can reclaim the device easily. Until the pairing secret can be
changed or another type of pairing is used this is the case.